requirements.txt
requests>=2.28.0
google-api-python-client>=2.70.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
python-dotenv>=0.21.0
